lets hide rpc in an abstraction for easier maintenance
